# backend/environments/python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Synergy API
  description: "A collaborative space for people who want to share their knowledge\
    \ and resource. Whether need to share in an organization or publically to all\
    \ or only to friends, Synergy is here. Anybody with need of getting some resources,\
    \ just search it and surely somebody will be there who has posted the related\
    \ content for you. This is a swagger documentation which provides API for Synergy\
    \ website."
  contact:
    email: cs1210547@iitd.ac.in
  version: "1.0"
servers:
- url: /
security:
- app_id: []
...
paths:
  /Post:
    put:
      tags:
      - Post
      summary: Update an existing Post
      description: Update an existing Post by Id
      operationId: update_post
      requestBody:
        description: Update an existent Post
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "400":
          description: Invalid ID supplied
        "404":
          description: Post not found
        "405":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.post_controller
    post:
      tags:
      - Post
      summary: Add a new Post to the database
      description: Add a new Post to the database
      operationId: add_post
      requestBody:
        description: Create a new Post in the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "405":
          description: Invalid input
      x-openapi-router-controller: swagger_server.controllers.post_controller
    delete:
      tags:
      - Post
      summary: Deletes Post from the database
      description: Deletes Post from the database
      operationId: delete_post
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Post not found
      x-openapi-router-controller: swagger_server.controllers.post_controller
  /Post/{Id}:
    get:
      tags:
      - Post
      summary: Find Post by ID
      description: For valid response use only the ID structure
      operationId: get_post_by_id
      parameters:
      - name: Id
        in: path
        description: Id of Post that needs to be fetched
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id.Post_Id'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "400":
          description: Invalid ID supplied
        "404":
          description: Account not found
      x-openapi-router-controller: swagger_server.controllers.post_controller
...
components:
  schemas:
    Post:
      required:
      - Author
      - Content
      - Id
      - Public_post
      - Report_num
      - Time
      - Title
      - Upvote_num
      type: object
      properties:
        Id:
          $ref: '#/components/schemas/Id.Post_Id'
        Author:
          $ref: '#/components/schemas/Id.Account_Id'
        Time:
          type: string
          format: date-time
        Title:
          type: string
        Content:
          type: string
        Comment:
          type: array
          items:
            $ref: '#/components/schemas/Id.Comment_Id'
        Public_post:
          type: boolean
        Institutes:
          type: array
          items:
            $ref: '#/components/schemas/Id.Institution_Id'
        tag_list:
          type: array
          items:
            $ref: '#/components/schemas/Id.Tag_Id'
        Report_num:
          type: integer
        Upvote_num:
          type: integer
      example:
        Comment:
        - obj: C
          uniq: uniq
        - obj: C
          uniq: uniq
        Public_post: true
        Upvote_num: 6
        Content: Content
        tag_list:
        - obj: T
          uniq: uniq
        - obj: T
          uniq: uniq
        Institutes:
        - obj: I
          uniq: uniq
        - obj: I
          uniq: uniq
        Author:
          obj: A
          uniq: uniq
        Title: Title
        Time: 2000-01-23T04:56:07.000+00:00
        Id:
          obj: P
          uniq: uniq
        Report_num: 0
...
    Id.Account_Id:
      required:
      - obj
      - uniq
      type: object
      properties:
        obj:
          type: string
          enum:
          - A
        uniq:
          type: string
      example:
        obj: A
        uniq: uniq
    Id.Institution_Id:
      required:
      - obj
      - uniq
      type: object
      properties:
        obj:
          type: string
          enum:
          - I
        uniq:
          type: string
      example:
        obj: I
        uniq: uniq
    Id.Tag_Id:
      required:
      - obj
      - uniq
      type: object
      properties:
        obj:
          type: string
          enum:
          - T
        uniq:
          type: string
      example:
        obj: T
        uniq: uniq
    Id.Comment_Id:
      required:
      - obj
      - uniq
      type: object
      properties:
        obj:
          type: string
          enum:
          - C
        uniq:
          type: string
      example:
        obj: C
        uniq: uniq
    Id.Post_Id:
      required:
      - obj
      - uniq
      type: object
      properties:
        obj:
          type: string
          enum:
          - P
        uniq:
          type: string
      example:
        obj: P
        uniq: uniq
  securitySchemes:
    app_id:
      type: apiKey
      description: API key to authorize requests
      name: appid
      in: query
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_app_id